requirements.txt
requests
bs4
beautifulsoup4
pillow
fpdf
PyPDF2